tests/requirements.txt
requests==2.31.0
pillow==10.0.0
selenium==4.11.2
pytest==7.4.0
pytest-html==3.2.0
pytest-xdist==3.3.1 